+++ CHANGELOG.md
@@ -220,5 +220,12 @@
 - Add SECURITY.md
 - Optimize lookup with binary search (PR)
 
+## 2026-09-01
+- Change Cargo.grvdep to Libs.grvdep as Gravel doesn't use Cargo
+
+## 2026-09-04
+- Add a `char` type
+- Start a Github wiki
+
 TODO: Add checker, imports and packages to documentation
 KNOWN BUGS: void functions return 0